minor changes to remove warnings. updated gitignore back to VSC

+++ my_lz/headers/lz.h
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <stdint.h>
 
-void find_best_match(char* sliding_window, char* lookahead, int position, int lookahead_filled, uint8_t *best_length, uint8_t *best_offset);
+void find_best_match(char* sliding_window, char* lookahead, int position, int lookahead_filled, int *best_length, int *best_offset);
 void compress_lzss(FILE *in, FILE *out);
 void decompress_lzss(FILE *in, FILE *out);
